requirements.txt
agate==1.7.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.3.0
dbt-core==1.7.3
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
debugpy==1.8.7
decorator==5.1.1
delta-spark==3.0.0
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
importlib-metadata==6.11.0
imutils==0.5.4
indian-names==0.3
ipython==8.28.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.0.2
leather==0.4.0
Logbook==1.5.3
MarkupSafe==3.0.1
mashumaro==3.13.1
matplotlib-inline==0.1.7
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.1.3
parsedatetime==2.6
parso==0.8.4
pathspec==0.11.2
prompt_toolkit==3.0.48
protobuf==5.28.2
pure_eval==0.2.3
py4j==0.10.9.7
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
spotipy==2.23.0
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
zipp==3.20.2
